Modified latitude and longitude to be Float

+++ x-guardian-core/app/models/Location.scala
@@ -1,3 +1,3 @@
 package models
 
-case class Location(latitude: Long, longitude: Long)
+case class Location(latitude: Float, longitude: Float)
